+++ charger/sgready.go
@@ -19,6 +19,7 @@ package charger
 
 import (
 	"context"
+	"errors"
 
 	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/charger/heating"
@@ -29,12 +30,21 @@ import (
 // SgReady charger implementation
 type SgReady struct {
 	*embed
-	*heating.SgReadyModeController
+	mode  int64
+	modeS func(int64) error
+	modeG func() (int64, error)
 }
 
 func init() {
 	registry.AddCtx("sgready", NewSgReadyFromConfig)
 }
+
+const (
+	_ int64 = iota
+	Normal
+	Boost
+	Stop
+)
 
 //go:generate go tool decorate -f decorateSgReady -b *SgReady -r api.Charger -t "api.Meter,CurrentPower,func() (float64, error)" -t "api.MeterEnergy,TotalEnergy,func() (float64, error)" -t "api.Battery,Soc,func() (float64, error)" -t "api.SocLimiter,GetLimitSoc,func() (int64, error)"
 
@@ -84,11 +94,51 @@ func NewSgReadyFromConfig(ctx context.Context, other map[string]interface{}) (ap
 // NewSgReady creates SG Ready charger
 func NewSgReady(ctx context.Context, embed *embed, modeS func(int64) error, modeG func() (int64, error)) (*SgReady, error) {
 	res := &SgReady{
-		embed:                 embed,
-		SgReadyModeController: heating.NewSgReadyModeController(ctx, modeS, modeG),
+		embed: embed,
+		mode:  Normal,
+		modeS: modeS,
+		modeG: modeG,
 	}
 
 	return res, nil
+}
+
+func (wb *SgReady) getMode() (int64, error) {
+	if wb.modeG == nil {
+		return wb.mode, nil
+	}
+	return wb.modeG()
+}
+
+// Status implements the api.Charger interface
+func (wb *SgReady) Status() (api.ChargeStatus, error) {
+	mode, err := wb.getMode()
+	if err != nil {
+		return api.StatusNone, err
+	}
+
+	if mode == Stop {
+		return api.StatusNone, errors.New("stop mode")
+	}
+
+	status := map[int64]api.ChargeStatus{Boost: api.StatusC, Normal: api.StatusB}
+	return status[mode], nil
+}
+
+// Enabled implements the api.Charger interface
+func (wb *SgReady) Enabled() (bool, error) {
+	mode, err := wb.getMode()
+	return mode == Boost, err
+}
+
+// Enable implements the api.Charger interface
+func (wb *SgReady) Enable(enable bool) error {
+	mode := map[bool]int64{false: Normal, true: Boost}[enable]
+	err := wb.modeS(mode)
+	if err == nil {
+		wb.mode = mode
+	}
+	return err
 }
 
 // MaxCurrent implements the api.Charger interface
