+++ charger/vaillant.go
@@ -26,7 +26,6 @@ import (
 
 	"github.com/WulfgarW/sensonet"
 	"github.com/evcc-io/evcc/api"
-	"github.com/evcc-io/evcc/charger/heating"
 	"github.com/evcc-io/evcc/util"
 	"github.com/evcc-io/evcc/util/request"
 	"github.com/samber/lo"
@@ -94,17 +93,17 @@ func NewVaillantFromConfig(ctx context.Context, other map[string]interface{}) (a
 	}
 
 	systemId := homes[0].SystemID
-	isHeating := cc.HeatingSetpoint > 0
+	heating := cc.HeatingSetpoint > 0
 
 	set := func(mode int64) error {
 		switch mode {
-		case heating.Normal:
-			if isHeating {
+		case Normal:
+			if heating {
 				return conn.StopZoneQuickVeto(systemId, cc.HeatingZone)
 			}
 			return conn.StopHotWaterBoost(systemId, sensonet.HOTWATERINDEX_DEFAULT)
-		case heating.Boost:
-			if isHeating {
+		case Boost:
+			if heating {
 				return conn.StartZoneQuickVeto(systemId, cc.HeatingZone, cc.HeatingSetpoint, 4) // hours
 			}
 			return conn.StartHotWaterBoost(systemId, sensonet.HOTWATERINDEX_DEFAULT) // zone 255
@@ -143,7 +142,7 @@ func NewVaillantFromConfig(ctx context.Context, other map[string]interface{}) (a
 	}
 
 	var heatingTempSensor bool
-	if isHeating {
+	if heating {
 		system, err := conn.GetSystem(systemId)
 		if err != nil {
 			return nil, err
@@ -152,7 +151,7 @@ func NewVaillantFromConfig(ctx context.Context, other map[string]interface{}) (a
 	}
 
 	var temp func() (float64, error)
-	if !isHeating || heatingTempSensor {
+	if !heating || heatingTempSensor {
 		temp = util.Cached(func() (float64, error) {
 			system, err := conn.GetSystem(systemId)
 			if err != nil {
